Added different aspect ratio support
- Added aspect ratio support
The data for the game itself does not change, X & Y of objects are independent of client aspect ratio. The rendering for the client is what allows different browser sizes to be supported. The ratio between the server's room sizes and client's bowser sizes are respected when calculating the rendering positions of object X's and Y's.